Updated the NPC to be more consistently interactable
For some reason you couldn't interact with the NPC consistently when they were standing still at the end of their walking, so I decided to make them pace back and forth to allow the player to continue to interact with them. This has made it so that the player can consistently interact with the NPC from the balcony for the whole game.