Added things to yards, pushing fab
Added some things to the backyards (a pool, a swing set, a slide, and a fence) to make them come to life a little bit. Made it so that the dog interaction spots are only interactable after seeing the outside dialogue, and also moved them into slightly better spots. Continuing to push the fab folder