Adding trigger based DT switches
On certain triggers, the data table for the NPC will now switch to a new one to allow new dialogue to occur based on that trigger. I wanted to keep it all in one data table but couldn't figure out how to swap rows of the data table from within a blueprint.