Added fire truck
Added a fire truck that will move to the players house. Needs to be changed to go at the end of the game, but more pressing is that the turn looks terrible. Arguably not a huge deal, but I'd like to fix that. Also modified some dialogue data tables, as well as fixed the game timer not properly sending the trigger to end the game at 0 seconds.